requirements.txt
Django~=5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

psycopg2-binary==2.9.9
environs~=11.0.0
drf-yasg==1.21.7
gunicorn==21.2.0

pillow~=10.3.0
beautifulsoup4==4.12.3
requests==2.31.0

PyJWT~=2.8.0
celery==5.3.6
redis==5.0.1

asgiref~=3.8.1
channels~=4.1.0
daphne==4.1.2
